requirements.txt
langchain
langchain-openai
langchain-community
duckduckgo-search
python-dotenv
langgraph
fastapi
uvicorn
